+++ R/plotMetric.R
@@ -9,6 +9,7 @@
 #' @param x the x-axis variable to plot
 #' @param imageId the ID of the image/fov
 #' @param ID the (optional) ID for plotting combinations
+#' @param ... Other parameters passed to `ggplot2` functions
 #'
 #' @return a `ggplot` object
 #' @export
@@ -31,33 +32,33 @@
 #' @import dplyr ggplot2
 #' @importFrom methods is
 plotMetricPerFov <- function(metricDf, theo = FALSE, correction = NULL,
-    x = NULL, imageId = NULL, ID = NULL) {
+    x = NULL, imageId = NULL, ID = NULL, ...) {
     # type checking
     stopifnot(is(metricDf, "data.frame"))
     stopifnot(is(correction, "character"))
     stopifnot(is(x, "character"))
     p <- ggplot(metricDf, aes(
         x = .data[[x]], y = .data[[correction]],
-        group = factor(.data[[imageId]])
+        group = factor(.data[[imageId]]), ...
     ))
     if (!is.null(ID)) {
         p <- p +
             geom_line(aes(colour = factor(.data[[ID]]))) +
-            facet_wrap(selection ~ ID)
+            facet_wrap(selection ~ ID, ...)
     } else {
         p <- p +
-            geom_line(aes(colour = factor(.data[[imageId]])))
+            geom_line(aes(colour = factor(.data[[imageId]])), ...)
     }
     p <- p +
         theme_minimal() +
-        theme(legend.position = "none") +
+        theme(legend.position = "none", ...) +
         labs(title = paste0(
             metricDf$fun, " metric for ",
             unique(metricDf$selection)
         ))
     if (theo == TRUE) {
         p <- p + geom_line(aes(x = .data[[x]], y = theo),
-            linetype = "dashed", color = "black"
+            linetype = "dashed", color = "black", ...
         )
     }
     return(p)
@@ -75,12 +76,13 @@ plotMetricPerFov <- function(metricDf, theo = FALSE, correction = NULL,
 #' @param x the x-axis variable to plot
 #' @param imageId the ID of the image/fov
 #' @param ID the (optional) ID for plotting combinations
+#' @param ... Other parameters passed to `ggplot2` functions
 #'
 #' @return a ggplot object
 #' @export
 #'
 #' @importFrom methods is
-plotCrossFOV <- function(subFov, theo, correction, x, imageId, ID = NULL) {
+plotCrossFOV <- function(subFov, theo, correction, x, imageId, ID = NULL, ...) {
     # type checking
     stopifnot(is(subFov, "data.frame"))
     #  Apply plot metric function for each combination
@@ -88,7 +90,7 @@ plotCrossFOV <- function(subFov, theo, correction, x, imageId, ID = NULL) {
         plotMetricPerFov(
             metricDf = subFov[subFov$selection == sel, ],
             theo = theo, correction = correction, x = x,
-            imageId = imageId, ID = ID
+            imageId = imageId, ID = ID, ...
         )
     })
     #  Count number of marks
@@ -113,6 +115,7 @@ plotCrossFOV <- function(subFov, theo, correction, x, imageId, ID = NULL) {
 #' @param x the x-axis variable to plot
 #' @param imageId the ID of the image/fov
 #' @param ID the (optional) ID for plotting combinations
+#' @param ... Other parameters passed to `ggplot2` functions
 #'
 #' @return a ggplot object
 #' @export
@@ -140,7 +143,8 @@ plotCrossMetricPerFov <- function(
         correction = NULL,
         x = NULL,
         imageId = NULL,
-        ID = NULL) {
+        ID = NULL,
+        ...) {
     # type checking
     stopifnot(is(metricDf, "data.frame"))
     stopifnot(is(imageId, "character"))
@@ -152,7 +156,7 @@ plotCrossMetricPerFov <- function(
         subFov <- metricDf[metricDf[[imageId]] %in% fov, ]
         return(plotCrossFOV(
             subFov = subFov, theo = theo, correction = correction,
-            x = x, imageId = imageId, ID = ID
+            x = x, imageId = imageId, ID = ID, ...
         ))
     })